Rename `sql` package to `sqlprojection`.
