--- sql/doc.go
@@ -1,2 +1,0 @@
-// Package sql provides utilities for building SQL-based projections.
-package sql
