+++ sqlprojection/doc.go
@@ -1,0 +1,2 @@
+// Package sqlprojection provides utilities for building SQL-based projections.
+package sqlprojection
