+++ sqlprojection/mysql/compat.go
