+++ sqlprojection/mysql/driver_test.go
@@ -5,8 +5,8 @@ import (
 	"database/sql"
 	"fmt"
 
-	"github.com/dogmatiq/projectionkit/sql/internal/drivertest"
-	. "github.com/dogmatiq/projectionkit/sql/mysql"
+	"github.com/dogmatiq/projectionkit/sqlprojection/internal/drivertest"
+	. "github.com/dogmatiq/projectionkit/sqlprojection/mysql"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
