+++ sqlprojection/sqlite/compat.go
