+++ sqlprojection/sqlite/error_cgo.go
