+++ sqlprojection/sqlite/error_go.go
